Add RuneLite 1.13 compatibility test suite
* Allow testing against RuneLite release and integration APIs

* Use long RuneLite item prices in loot telemetry

* Use long RuneLite item prices in notifications

* Add RuneLite 1.13 compatibility regression suite

* Test every change against RuneLite release and integration APIs

+++ src/main/java/com/sixthdegree/SixthDegreeLootService.java
@@ -88,7 +88,7 @@ final class SixthDegreeLootService
 	{
 		try
 		{
-			int price = itemManager.getItemPrice(itemId);
+			long price = itemManager.getItemPrice(itemId);
 			if (price > 0)
 			{
 				return price;
